toolkit/spatial-navigation/tests/chrome/test_snav_textFields.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css type="text/css"?>

<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=436084
-->

<window title="Mozilla Bug 288254"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
  onload="onLoad();">

  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      

<body id="some-content" xmlns="http://www.w3.org/1999/xhtml">

<p>
<a id="start" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 436084</a>
</p>

<p>
<input id="textinput" value="abcdefghijklmnopqrstuvwxyz" type="text"></input>
</p>

<p>
<textarea id="textarea" name="textarea" cols="30" rows="3" wrap="hard">The Book of Mozilla</textarea>
</p>

<p>
<a id="end" href="https://bugzilla.mozilla.org/show_bug.cgi?id=436084">Mozilla Bug 436084</a>
</p>

</body>


<script class="testbody" type="application/javascript">
<![CDATA[
Components.utils.import("resource://gre/modules/SpatialNavigation.js");

function onLoad()
{
    var x = document.getElementById("some-content");
    var snav = new SpatialNavigation(x);

    function moveAndVerify(direction, value)
    {
        sendKey(direction, document.activeElement);
        ok(document.activeElement.getAttribute("id") == value, "Move");
    }
    
    // get to a known place.
    document.getElementById("start").focus();
    
    // from start.
    moveAndVerify("DOWN", "textinput");
    moveAndVerify("DOWN", "textinput");

    moveAndVerify("DOWN", "textarea");
    moveAndVerify("DOWN", "textarea");

    moveAndVerify("DOWN", "end");

    moveAndVerify("UP", "textarea");
    moveAndVerify("UP", "textarea");

    moveAndVerify("UP", "textinput");
    moveAndVerify("UP", "textinput");

    moveAndVerify("UP", "start");

    moveAndVerify("DOWN", "textinput");

    moveAndVerify("LEFT", "textinput");

    for (var i=26; i>0; i--)
        moveAndVerify("RIGHT", "textinput")

    moveAndVerify("RIGHT", "textarea")    


    moveAndVerify("LEFT", "textarea");

    // move 19 right
    for (var i=19; i>0; i--)
        moveAndVerify("RIGHT", "textarea");

    moveAndVerify("RIGHT", "end");

    moveAndVerify("UP", "textarea"); 

    sendString("I like to eat apples and bananas", document.activeElement);

    SimpleTest.finish();
}

]]></script>
</window>
